+++ config/routes.rb
@@ -124,7 +124,7 @@ Rails.application.routes.draw do
               end
               resource :bulk, only: [:create, :update, :destroy], controller: :bulk
               resources :filters, only: :index
-              resources :people, only: [:index, :update]
+              resources :people, only: [:index, :show, :update, :destroy]
             end
           end
         end